Remove a LOT of C++
