--- .gitmodules
@@ -1,4 +1,0 @@
-[submodule "opencv-build"]
-	path = opencv-build
-	url = https://github.com/lasarobotics/opencv-builds
-	branch = master
