--- ftc-cameratest-java/build.gradle
@@ -1,33 +1,0 @@
-apply plugin: 'com.android.application'
-
-android {
-    compileSdkVersion 21
-    buildToolsVersion "21.1.2"
-
-    defaultConfig {
-        applicationId "com.lasarobotics.tests.camera"
-        minSdkVersion 16
-        targetSdkVersion 16
-    }
-
-    buildTypes {
-        debug {
-            jniDebuggable true
-        }
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.txt'
-        }
-    }
-}
-
-repositories {
-    flatDir {
-        dirs 'libs'
-    }
-}
-
-dependencies {
-    compile project(':opencv-java')
-    compile project(':ftc-visionlib-java')
-}
