+++ ftc-cameratest/build.gradle
@@ -29,5 +29,5 @@ repositories {
 
 dependencies {
     compile project(':opencv-java')
-    compile project(':ftc-visionlib')
+    compile project(':ftc-visionlib-java')
 }
