--- ftc-jnibuild/build.gradle
@@ -1,5 +1,0 @@
-apply plugin: 'java'
-
-dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-}
