+++ settings.gradle
@@ -1,10 +1,7 @@
 //Libraries
 include ':opencv-java'
 include ':ftc-visionlib-java'
-//include ':ftc-jnibuild'
-//include ':ftc-visionlib'
 
 //Demos and Tests
-//include ':ftc-cameratest'
-include ':ftc-cameratest-java'
+include ':ftc-cameratest'
 //include ':ftc-robotcontroller'
